+++ modules/3d/src/ptcloud/sac_segmentation.cpp
@@ -29,7 +29,7 @@ SACSegmentation::segmentSingle(Mat &points, std::vector<bool> &label, Mat &model
     double _threshold = threshold * threshold;
     int state = (int) rng_state;
     const int points_size = points.rows * points.cols / 3;
-    const double _radius_min = radius_min, _radius_max = radius_min;
+    const double _radius_min = radius_min, _radius_max = radius_max;
     const Ptr <ModelConstraintFunction> _custom_constraint = custom_model_constraints;
     Ptr <ModelConstraintFunction> constraint_func = custom_model_constraints;
 
@@ -76,7 +76,7 @@ SACSegmentation::segmentSingle(Mat &points, std::vector<bool> &label, Mat &model
             constraint_func = makePtr<ModelConstraintFunction>(
                     [_radius_min, _radius_max, _custom_constraint]
                             (const std::vector<double> &model_coeffs) {
-                        double radius = model_coeffs[4];
+                        double radius = model_coeffs[3];
                         return radius >= _radius_min && radius <= _radius_max &&
                                (_custom_constraint.empty() || (*_custom_constraint)(model_coeffs));
                     }
