--- modules/3d/test/test_segmentation.cpp
@@ -1,284 +1,0 @@
-// This file is part of OpenCV project.
-// It is subject to the license terms in the LICENSE file found in the top-level directory
-// of this distribution and at http://opencv.org/license.html.
-
-#include "test_precomp.hpp"
-
-namespace opencv_test { namespace {
-class PlaneSegmentationTest : public ::testing::Test
-{
-public:
-    vector<vector<float>> models = {
-            {0, 0, 1, 0},
-            {1, 0, 0, 0},
-            {0, 1, 0, 0},
-            {1, 1, 1, -150},
-    };
-    vector<float> thrs = {0.1f, 0.2f, 0.4f, 0.5f};
-    vector<int> pt_nums = {100, 200, 300, 400};
-    vector<vector<float>> limits = {
-            {5,  55, 5,  55, 0, 0},
-            {0,  0,  5,  55, 5, 55},
-            {5,  55, 0,  0,  5, 55},
-            {10, 50, 10, 50, 0, 0},
-    };
-
-    int models_num = (int) models.size();
-    // Used to store point cloud, generated plane and model
-    Mat pt_cloud, generated_pts, detected_plane_models;
-    vector<int> label;
-    SACSegmentation sacSegmentation;
-
-    void SetUp() override
-    {
-        // Set basic sac arguments.
-        sacSegmentation.setSacModelType(SAC_MODEL_PLANE);
-        sacSegmentation.setSacMethodType(SAC_METHOD_RANSAC);
-        sacSegmentation.setConfidence(1);
-    }
-
-    /**
-     * Used to generate a specific plane with random points
-     * model: plane coefficient [a,b,c,d] means ax+by+cz+d=0
-     * thr: generate the maximum distance from the point to the plane
-     * limit: the range of xyz coordinates of the generated plane
-     **/
-    static void generatePlane(Mat &pt_cloud, const vector<float> &model, float thr, int num,
-                              const vector<float> &limit)
-    {
-        pt_cloud = Mat(num, 3, CV_32F);
-        cv::RNG rng(0);
-        auto *pt_cloud_ptr = (float *) pt_cloud.data;
-
-        // Part of the points are generated for the specific model.
-        // The other part of the points are used to increase the thickness of the plane.
-        int std_num = (int) num / 3;
-        // Difference of maximum d between two parallel planes.
-        float d_thr = thr * sqrt(model[0] * model[0] + model[1] * model[1] + model[2] * model[2]);
-
-        for (int i = 0; i < num; i++)
-        {
-            // Let d change then generate thickness.
-            float d = i < std_num ? model[3] : rng.uniform(model[3] - d_thr, model[3] + d_thr);
-            float x, y, z;
-            // c is 0 means the plane is vertical.
-            if (model[2] == 0)
-            {
-                z = rng.uniform(limit[4], limit[5]);
-                if (model[0] == 0)
-                {
-                    x = rng.uniform(limit[0], limit[1]);
-                    y = -d / model[1];
-                }
-                else if (model[1] == 0)
-                {
-                    x = -d / model[0];
-                    y = rng.uniform(limit[2], limit[3]);
-                }
-                else
-                {
-                    x = rng.uniform(limit[0], limit[1]);
-                    y = -(model[0] * x + d) / model[1];
-                }
-            }
-                // c is not 0.
-            else
-            {
-                x = rng.uniform(limit[0], limit[1]);
-                y = rng.uniform(limit[2], limit[3]);
-                z = -(model[0] * x + model[1] * y + d) / model[2];
-            }
-
-            pt_cloud_ptr[3 * i] = x;
-            pt_cloud_ptr[3 * i + 1] = y;
-            pt_cloud_ptr[3 * i + 2] = z;
-        }
-    }
-};
-
-void normalizeMat(Mat &m)
-{
-    m.convertTo(m, CV_32F);
-    float l = (float) sqrt(m.dot(m));
-    m /= l;
-}
-
-float getDiff(const Mat &a, const Mat &b)
-{
-    // If the direction is the same
-    Mat d1 = a - b;
-    // If the direction is not the same
-    Mat d2 = a + b;
-    return (float) min(d1.dot(d1), d2.dot(d2));
-}
-
-int countNum(const vector<int> &m, int num)
-{
-    int t = 0;
-    for (int a: m)
-        if (a == num) t++;
-    return t;
-}
-
-TEST_F(PlaneSegmentationTest, SinglePlaneDetection)
-{
-    sacSegmentation.setMaxIterations(500);
-    sacSegmentation.setNumberOfModelsExpected(1);
-
-    // Single plane detection
-    for (int i = 0; i < models_num; i++)
-    {
-        generatePlane(generated_pts, models[i], thrs[i], pt_nums[i], limits[i]);
-        pt_cloud.push_back(generated_pts);
-        //A point with a distance equal to the threshold is not considered an inliner point
-        sacSegmentation.setDistanceThreshold(thrs[i] + 0.01);
-        int num = sacSegmentation.segment(pt_cloud, label, detected_plane_models);
-
-        ASSERT_EQ(1, num)
-        << "Model number should be equal to 1.";
-        ASSERT_EQ(pt_cloud.rows, label.size())
-        << "Label size should be equal to point number.";
-
-        Mat ans_model(1, 4, CV_32F, models[i].data()), detected_model(detected_plane_models);
-        normalizeMat(ans_model);
-        normalizeMat(detected_model);
-
-        EXPECT_LE(getDiff(ans_model, detected_model), 0.0001)
-        << "Initial model is " << ans_model << ". Detected model is " << detected_model
-        << ". The difference in coefficients should not be too large.";
-        ASSERT_EQ(pt_nums[i], countNum(label, 1))
-        << "There are " << pt_nums[i] << " points need to be marked.";
-
-        for (int j = 0; j < pt_cloud.rows; j++)
-        {
-            if (j < pt_cloud.rows - pt_nums[i])
-                ASSERT_EQ(0, label[j])
-                << "This index should not be marked: " << j << ". This point is "
-                << pt_cloud.row(j);
-            else
-                ASSERT_EQ(1, label[j])
-                << "This index should be marked: " << j << ". This point is "
-                << pt_cloud.row(j);
-        }
-    }
-}
-
-TEST_F(PlaneSegmentationTest, MultiplePlaneDetection)
-{
-    sacSegmentation.setMaxIterations(1000);
-    sacSegmentation.setDistanceThreshold(0.51);
-    sacSegmentation.setNumberOfModelsExpected(models_num);
-
-    for (int i = 0; i < models_num; i++)
-    {
-        generatePlane(generated_pts, models[i], 0.5, pt_nums[i], limits[i]);
-        pt_cloud.push_back(generated_pts);
-    }
-
-    // Multiple planes detection
-    int num = sacSegmentation.segment(pt_cloud, label, detected_plane_models);
-
-    ASSERT_EQ(models_num, num)
-    << "Model number should be equal to " << models_num << ".";
-    ASSERT_EQ(pt_cloud.rows, label.size())
-    << "Label size should be equal to point number.";
-
-    int checked_num = 0;
-    for (int i = 0; i < models_num; i++)
-    {
-        Mat ans_model(1, 4, CV_32F, models[models_num - 1 - i].data()), detected_model(
-                detected_plane_models.row(i));
-        normalizeMat(ans_model);
-        normalizeMat(detected_model);
-
-        EXPECT_LE(getDiff(ans_model, detected_model), 0.0001)
-        << "Initial model is " << ans_model << ". Detected model is " << detected_model
-        << ". The difference in coefficients should not be too large.";
-        ASSERT_EQ(pt_nums[models_num - 1 - i], countNum(label, i + 1))
-        << "There are " << pt_nums[i] << " points need to be marked.";
-
-        for (int j = checked_num; j < pt_nums[i]; j++)
-            ASSERT_EQ(models_num - i, label[j])
-            << "This index " << j << " should be marked as "<< models_num - i
-            << ". This point is " << pt_cloud.row(j);
-        checked_num += pt_nums[i];
-    }
-}
-
-TEST_F(PlaneSegmentationTest, PlaneDetectionWithConstraints)
-{
-    sacSegmentation.setMaxIterations(5000);
-    sacSegmentation.setDistanceThreshold(0.51);
-    sacSegmentation.setNumberOfModelsExpected(1);
-
-    for (int i = 0; i < models_num; i++)
-    {
-        generatePlane(generated_pts, models[i], 0.5f, pt_nums[i], limits[i]);
-        pt_cloud.push_back(generated_pts);
-    }
-
-    int checked_num = 0;
-    for (int i = 0; i < models_num; i++)
-    {
-
-        // The angle between the model normals and the constraints must be less than 1 degree
-        float radian_thr = 1 * 3.1415926f / 180;
-        vector<float> normal_constraint = {models[i][0], models[i][1], models[i][2]};
-
-        // Normal vector constraint function
-        ModelConstraintFunction constraint = [normal_constraint, radian_thr](
-                const vector<double> &model) -> bool {
-            vector<float> model_normal = {(float) model[0], (float) model[1], (float) model[2]};
-            float dot12 = normal_constraint[0] * model_normal[0] +
-                          normal_constraint[1] * model_normal[1] +
-                          normal_constraint[2] * model_normal[2];
-            float m1m1 = normal_constraint[0] * normal_constraint[0] +
-                         normal_constraint[1] * normal_constraint[1] +
-                         normal_constraint[2] * normal_constraint[2];
-            float m2m2 = model_normal[0] * model_normal[0] +
-                         model_normal[1] * model_normal[1] +
-                         model_normal[2] * model_normal[2];
-            float square_cos_theta = dot12 * dot12 / (m1m1 * m2m2);
-
-            return square_cos_theta >= cos(radian_thr) * cos(radian_thr);
-        };
-
-        // Make a pointer to constraint function
-        Ptr<ModelConstraintFunction> constraint_ptr = makePtr<ModelConstraintFunction>(constraint);
-        sacSegmentation.setCustomModelConstraints(constraint_ptr);
-
-        // Single plane detection with constraint
-        int num = sacSegmentation.segment(pt_cloud, label, detected_plane_models);
-
-        ASSERT_EQ(1, num)
-        << "Model number should be equal to " << models_num << ".";
-        ASSERT_EQ(pt_cloud.rows, label.size())
-        << "Label size should be equal to point number.";
-
-        Mat ans_model(1, 4, CV_32F, models[i].data()), detected_model(detected_plane_models);
-        normalizeMat(ans_model);
-        normalizeMat(detected_model);
-
-        EXPECT_LE(getDiff(ans_model, detected_model), 0.0001)
-        << "Initial model is " << ans_model << ". Detected model is " << detected_model
-        << ". The difference in coefficients should not be too large.";
-        ASSERT_EQ(pt_nums[i], countNum(label, 1))
-        << "There are " << pt_nums[i] << " points need to be marked.";
-
-        for (int j = 0; j < pt_cloud.rows; j++)
-        {
-            if (j >= checked_num && j < checked_num + pt_nums[i])
-                ASSERT_EQ(1, label[j])
-                << "This index should be marked: " << j << ". This point is "
-                << pt_cloud.row(j);
-            else
-                ASSERT_EQ(0, label[j])
-                << "This index should not be marked: " << j << ". This point is "
-                << pt_cloud.row(j);
-        }
-        checked_num += pt_nums[i];
-    }
-}
-
-} // namespace
-} // opencv_test
